@@ -39,7 +39,7 @@
     "dedent": "^0.7.0",
     "is-ci-cli": "^2.2.0",
     "jest": "^27.3.1",
-    "typescript": "^4.4.4"
+    "typescript": "^4.9.5"
   },
   "publishConfig": {
     "access": "public"
