@@ -1,7 +1,7 @@
 import { promises as fs } from 'fs'
 import rebuild, { MATCHERS_FILE_PATH } from '../_utils/rebuild'
 
-export default async function main(): Promise<number> {
+export async function main(): Promise<number> {
   await fs.writeFile(MATCHERS_FILE_PATH, await rebuild(), 'utf8')
 
   return 0
