@@ -1,19 +1,20 @@
-#!/usr/bin/env ts-node
-
 import * as t from '@babel/types'
 import { join } from 'path'
+import { BUILDER_KEYS, NodeField, NODE_FIELDS } from '../../src/NodeTypes'
+import format from './format'
 import {
-  stringifyValidator,
   isValidatorOfType,
+  stringifyType,
+  stringifyValidator,
   toFunctionName,
   typeForValidator,
-  stringifyType,
 } from './utils'
 import dedent = require('dedent')
-import { NodeField, BUILDER_KEYS, NODE_FIELDS } from '../../src/NodeTypes'
-import format from './format'
 
-export const MATCHERS_FILE_PATH = join(__dirname, '../../src/matchers.ts')
+export const MATCHERS_FILE_PATH = join(
+  __dirname,
+  '../../src/matchers/generated.ts'
+)
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 const toBindingIdentifierName: (name: string) => string = (t as any)
@@ -79,33 +80,14 @@ function writeToString(write: (writer: SimpleWriter) => void): string {
 }
 
 function rebuildTo(out: SimpleWriter): string | void {
+  out.write(`/* DO NOT EDIT. This file was generated by 'script/rebuild' */\n`)
   out.write(`/* eslint-disable */\n`)
   out.write(`import * as t from '@babel/types';\n\n`)
 
   out.write(dedent`
-    export * from './matchers/spacers';
-    export { default as Function } from './matchers/Function';
-    export { default as Matcher } from './matchers/Matcher';
-    export { default as anyExpression } from './matchers/anyExpression';
-    export { default as anyList } from './matchers/anyList';
-    export { default as anyNode } from './matchers/anyNode';
-    export { default as anyNumber } from './matchers/anyNumber';
-    export { default as anyStatement } from './matchers/anyStatement';
-    export { default as anyString } from './matchers/anyString';
-    export { default as anything } from './matchers/anything';
-    export { default as arrayOf } from './matchers/arrayOf';
-    export { default as capture, CaptureBase, CapturedMatcher } from './matchers/capture';
-    export { default as containerOf } from './matchers/containerOf';
-    export { default as fromCapture } from './matchers/fromCapture';
-    export { default as function } from './matchers/Function';
-    export { default as oneOf } from './matchers/oneOf';
-    export { default as or } from './matchers/or';
-    export { default as matcher } from './matchers/predicate';
-    export { default as tupleOf } from './matchers/tupleOf';
-
-    import tupleOf from './matchers/tupleOf';
-    import { isNode } from './NodeTypes';
-    import Matcher from './matchers/Matcher';
+    import { tupleOf } from './tupleOf';
+    import { Matcher } from './Matcher';
+    import { isNode } from '../NodeTypes';
   `)
 
   out.write('\n\n')
