@@ -1,3 +1,3 @@
 export * from './matchers'
-export { default as match } from './utils/match'
-export { default as matchPath } from './utils/matchPath'
+export { match } from './utils/match'
+export { matchPath } from './utils/matchPath'
