@@ -1,6 +1,6 @@
 import * as t from '@babel/types'
 import { isNode } from '../NodeTypes'
-import Matcher from './Matcher'
+import { Matcher } from './Matcher'
 
 export class AnyNodeMatcher extends Matcher<t.Node> {
   matchValue(value: unknown): value is t.Node {
@@ -8,6 +8,6 @@ export class AnyNodeMatcher extends Matcher<t.Node> {
   }
 }
 
-export default function anyNode(): Matcher<t.Node> {
+export function anyNode(): Matcher<t.Node> {
   return new AnyNodeMatcher()
 }
