@@ -1,33 +1,10 @@
+/* DO NOT EDIT. This file was generated by 'script/rebuild' */
 /* eslint-disable */
 import * as t from '@babel/types'
 
-export * from './matchers/spacers'
-export { default as Function } from './matchers/Function'
-export { default as Matcher } from './matchers/Matcher'
-export { default as anyExpression } from './matchers/anyExpression'
-export { default as anyList } from './matchers/anyList'
-export { default as anyNode } from './matchers/anyNode'
-export { default as anyNumber } from './matchers/anyNumber'
-export { default as anyStatement } from './matchers/anyStatement'
-export { default as anyString } from './matchers/anyString'
-export { default as anything } from './matchers/anything'
-export { default as arrayOf } from './matchers/arrayOf'
-export {
-  default as capture,
-  CaptureBase,
-  CapturedMatcher,
-} from './matchers/capture'
-export { default as containerOf } from './matchers/containerOf'
-export { default as fromCapture } from './matchers/fromCapture'
-export { default as function } from './matchers/Function'
-export { default as oneOf } from './matchers/oneOf'
-export { default as or } from './matchers/or'
-export { default as matcher } from './matchers/predicate'
-export { default as tupleOf } from './matchers/tupleOf'
-
-import tupleOf from './matchers/tupleOf'
-import { isNode } from './NodeTypes'
-import Matcher from './matchers/Matcher'
+import { tupleOf } from './tupleOf'
+import { Matcher } from './Matcher'
+import { isNode } from '../NodeTypes'
 
 // aliases for keyword-named functions
 export { Import as import }
